Remove light and group auto-detect and add to configuration.
Auto-detecting if the fan has a light and the fan's group would frequently fail with a network error. A network error during discovery resulted in the fan not showing up in HA. I hoping that discovery will go smoother now the user has to specify these in advance in a configuration file.